Add constructors that accept a custom endpoint
Add constructor variants that accept a custom AWS endpoint for use in
specifying a local endpoint during testing (e.g. using minio or
localstack). This could be achieved previously by creating a custom
session and passing that into `NewClientWithSession(...)`, but adding
the convenience constuctors makes it clearer how to approach this.